Update dependencies and refactor keyboard shortcuts management
- Bumped versions of several Nx packages to 21.3.1 for improved stability.
- Updated the keyboard shortcuts management system, including the addition of a new settings route for keyboard shortcuts.
- Refactored keyboard shortcut handling to enhance input validation and error logging.
- Removed deprecated keyboard shortcuts component and cleaned up related tests.
- Improved overall code organization and consistency across components.
