+++ app/src/features/keyboard/components/keyboard-shortcuts-manager.test.tsx
