+++ app/src/routes/(chat)/layout.tsx
@@ -1,11 +1,11 @@
 import { createFileRoute, Outlet } from "@tanstack/react-router";
 import { Authenticated, AuthLoading, Unauthenticated } from "convex/react";
 
-import { KeyboardShortcutsManager } from "@/components/keyboard-shortcuts-manager";
+import { KeyboardShortcutsManager } from "@/features/keyboard/components/keyboard-shortcuts-manager";
 import { ProgrammableSidebarProvider } from "@/components/programmable-sidebar-provider";
 import { AutoGuestSignIn } from "@/features/auth/components/auto-guest-signin";
 import { AiModelProvider } from "@/features/chat/providers/ai-model-provider";
-import useKeyboardShortcutHandler from "@/hooks/use-keyboard-shortcut-handler";
+import useKeyboardShortcutHandler from "@/features/keyboard/hooks/use-keyboard-shortcut-handler";
 
 const defaultOpen = ["left"];
 const keyboardShortcuts = { left: "b", right: "l" };
