+++ app/src/routes/(files)/layout.tsx
@@ -6,14 +6,14 @@ import { Authenticated } from "convex/react";
 import { Building, ChartArea, Key, Keyboard, MessageSquare, Shield, ToggleLeft, User, Users, Zap } from "lucide-react";
 import { Fragment } from "react";
 
-import { KeyboardShortcutsManager } from "@/components/keyboard-shortcuts-manager";
+import { KeyboardShortcutsManager } from "@/features/keyboard/components/keyboard-shortcuts-manager";
 import { ProgrammableSidebarProvider } from "@/components/programmable-sidebar-provider";
 import { useAuth } from "@/features/auth/lib/auth-ui-provider";
 import AppSidebar from "@/features/layout/components/app-sidebar";
 import type { NavItem } from "@/features/layout/components/nav-items";
 import { NavItems } from "@/features/layout/components/nav-items";
 import SiteHeader from "@/features/layout/components/site-header";
-import useKeyboardShortcutHandler from "@/hooks/use-keyboard-shortcut-handler";
+import useKeyboardShortcutHandler from "@/features/keyboard/hooks/use-keyboard-shortcut-handler";
 import { getAuthRedirectUrl } from "@/lib/utils";
 
 const sidebarHeader = (
